Обновил типы данных в конструкторах

+++ Runtime/Components/FinalBonus.cs
@@ -2,7 +2,7 @@ namespace LittleBit.Modules.Description.Components
 {
     public class FinalBonus : BaseAttribute
     {
-        public FinalBonus(float baseValue, float baseMultiplier) : base(baseValue, baseMultiplier)
+        public FinalBonus(double baseValue, double baseMultiplier) : base(baseValue, baseMultiplier)
         {
             
         }
